+++ CHANGELOG.md
@@ -17,3 +17,7 @@
 - 新增`Publish image`workflows，用于发布镜像
 - 新增`no-toxic-comments.yml`，用于检测评论
 - 更新`auto-close-pr.yml`，用于修正PR的目标到`develop`
+## 0.0.6
+- 更新docker镜像的构建
+- 【#12】新增配置`Notification:IsSingleAccountSingleNotify`，支持开启每个账号单独推送消息
+- publish-release.yml新增手动输入tag功能